Let's move Repo to Grid space as it's related
Repo Post was coupled with Grid system. In reality, it relies on the Factory
builder in order to return a proper collection. A repository's responsibility is
that it should retrieve data from a persistence layer and then build a
collection and/or object base on that. The Repo/Post was doing this specifically
for Grid, not just a general Repo. This may evolve in the future if requirements
change.